add: UserAction data fetch
1. setupProxy.js: logs기록도 함께 우회하기 위해 api 경로수정
2. userActionCardList: data fetch and render
3. userLog: data parsing 과정에서 useState로 상태관리 필요할 듯
4. 경과시간 로직 수정 필요
5. log 콘텐츠 font-weight 조절 필요
6. log api action 순서 조정 및 형태 변경 필요할 수도?...

+++ fe/todo-list/src/components/card/actionCardList.jsx
@@ -1,19 +1,45 @@
-import React from 'react';
+import axios from 'axios';
+import React, { useEffect, useState } from 'react';
 import styled from 'styled-components';
 import Icon from '../icon/icon';
 import ActionCard from './actionCard';
 
 const ActionCardList = ({ state, setState }) => {
+  const [data, setData] = useState(null);
+  const [loading, setLoading] = useState(false);
+  const [error, setError] = useState(null);
+
+  useEffect(() => {
+    const fetchData = async () => {
+      try {
+        setData(null);
+        setError(null);
+        setLoading(true);
+        const request = `/api/logs`;
+        const response = await axios.get(request);
+        setData(response.data);
+      } catch (error) {
+        setError(error);
+      }
+      setLoading(false);
+    };
+    fetchData();
+  }, []);
+
+  if (loading) return <div>Loading...</div>;
+  if (error) return <div>Error!!!</div>;
+  if (!data) return null;
+
+  console.log(data);
   return (
     <ActionCardContainer state={state}>
       <IconPosition onClick={setState}>
         <Icon type="delete" />
       </IconPosition>
       <ActionCardListBox>
-        <ActionCard />
-        <ActionCard />
-        <ActionCard />
-        <ActionCard />
+        {data.map((el) => (
+          <ActionCard key={el.id} data={el} />
+        ))}
       </ActionCardListBox>
     </ActionCardContainer>
   );
